Remove unnecessary casts and warn about them

+++ leshan-client-core/src/main/java/org/eclipse/leshan/client/resource/SimpleInstanceEnabler.java
@@ -161,7 +161,7 @@ public class SimpleInstanceEnabler extends BaseInstanceEnabler {
     }
 
     protected double createDefaultFloatValue(ObjectModel objectModel, ResourceModel resourceModel) {
-        return (double) Math.random() * 100;
+        return Math.random() * 100;
     }
 
     protected byte[] createDefaultOpaqueValue(ObjectModel objectModel, ResourceModel resourceModel) {
